fix: change error detection to simply log
Previous behavior treated any output on stderr a failure and would exit
the script. This made it difficult to debug as some _warnings_ were
output on stderr and trying to allow list all of them was not working.

The new behavior is that stderr is simply echoed back out to the console
so errors can been observed and addressed.

As such the `firstTime` run was removed since that was intended to
address the warnings but is no longer needed and did not succeed on my
system anyway.